Feat (pages) added review dynamic route
